Organize the dataTree folder
