+++ src/components/dataTree/Node.module.scss
@@ -1,4 +1,4 @@
-.node {
+.domNode {
   margin-left: 4px;
 }
